bizq-domain-suggestions/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-whois==0.8.0
openai==1.3.7
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
structlog==23.2.0
pydantic-settings==2.1.0